+++ plone/app/contenttypes/content.py
@@ -18,10 +18,11 @@ from plone.dexterity.content import Container
 from Products.CMFCore.utils import getToolByName
 
 from zope.interface import implements
+from Products.CMFPlone.interfaces.syndication import ISyndicatable
 
 
 class Collection(Item):
-    implements(ICollection)
+    implements(ICollection, ISyndicatable)
 
     #security.declareProtected(View, 'listMetaDataFields')
     def listMetaDataFields(self, exclude=True):